mysite/requirement.txt
asgiref==3.4.1
Django==3.2.6
django-tinymce4-lite==1.8.0
jsmin==2.2.2
numpy==1.21.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
